// src/model/classes/viatics/ExpenseAdvanceDefaultRQ.tsx
export class ExpenseAdvanceDefaultRQ {

    ProcessType: string;
    International: boolean;
    Destination: string;
    IDUser: number;
    NumberOfDays: number;
    StartTime: number;
    EndTime: number;
    FamilyHome: boolean;
    PaymentExpenses: boolean;
    PaymentDestination: string;
    CalculateAdvances: boolean;
    countFamilyHome: number;
    countPaymentDestination: number;
    numDaysMultiDest: number;
    HasFlight: boolean;
    HasTransport: boolean;
    HasHotel: boolean;
}